feat: update MongoDB.Driver to 3.0.0
In 3.0.0, all of Core has been merged into one, which is why we can drop dependency on `MongoDB.Driver.GridFS`.

`GuidRepresentation` has also been altered, so now to get a default Guid representation behaviour, we must register a global `GuidSerializer` with a selected representation.

Alternatively, a representation must be choosen on a per-field basis using annotations.