feat: PoC assuming client uses `CSharpLegacy`
This is not intended for production - but PoC to see if I can get test suite green.

Here I just assume all clients will be globally using `CSharpLegacy` representation.

In reality, we need to address the now explicit choice of Guid representation that MongoDB.Driver enforces, which allows Guid's to be represented differently, even within the same document.